remove the mapper function and instead make it an extension function inside the ANime Entity class
