+++ Abhi-AnimeX/app/src/main/java/com/abhijith/animex/domain/model/AnimeItem.kt
@@ -1,5 +1,7 @@
 package com.abhijith.animex.domain.model
 
+import com.abhijith.animex.data.model.AnimeEntity
+
 data class AnimeItem(
     val title: String,
     val source: String,
@@ -13,3 +15,20 @@ data class AnimeItem(
     val imageUrl: String,
     val youtubeId: String
 )
+
+
+fun AnimeEntity.toDomain(): AnimeItem {
+    return AnimeItem(
+        title = title,
+        source = source,
+        year = year,
+        japaneseName = japaneseName,
+        rating = rating,
+        rank = rank,
+        score = score,
+        synopsis = synopsis,
+        genres = genres,
+        imageUrl = imageUrl,
+        youtubeId = youtubeId
+    )
+}
